+++ boot/src/asm/port.rs
@@ -1,0 +1,36 @@
+//! # Wrapper functions of x64 instructions
+//! ## References
+//! * [Intel 64 and IA-32 Architectures Software Developer's Manual December 2023](https://www.intel.com/content/www/us/en/developer/articles/technical/intel-sdm.html)
+
+use core::arch::asm;
+
+/// # Input from port
+/// ## References
+/// * [Intel 64 and IA-32 Architectures Software Developer's Manual December 2023](https://www.intel.com/content/www/us/en/developer/articles/technical/intel-sdm.html) Vol.2A 3-505
+#[inline(never)]
+pub fn inb(port: u16) -> u8 {
+    let mut data: u8;
+    unsafe {
+        asm!(
+            "in al, dx",
+            in("dx") port,
+            out("al") data,
+        );
+    }
+    data
+}
+
+/// # Output to port
+/// ## References
+/// * [Intel 64 and IA-32 Architectures Software Developer's Manual December 2023](https://www.intel.com/content/www/us/en/developer/articles/technical/intel-sdm.html) Vol.2B 4-176
+#[inline(never)]
+pub fn outb(port: u16, data: u8) {
+    unsafe {
+        asm!(
+            "out dx, al",
+            in("dx") port,
+            in("al") data,
+        );
+    }
+}
+
